# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "contact": {
      "name": "NIEMOpen",
      "url": "https://niemopen.org"
    },
    "description": "Community support for common NIEM tool functionality to browse and search the model, build and migrate subsets, and check conformance",
    "title": "NIEM API 2.0",
    "version": "2.0-alpha07:SNAPSHOT"
  },
  "servers": [
    {
      "url": "https://tools.niem.gov/api/v2",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080/api/v2",
      "description": "Local machine"
    },
    {
      "url": "https://yellowkiwi06.icl.gtri.org/",
      "description": "Internal test server"
    }
  ],
...
  "paths": {
...
    "/stewards/{stewardKey}": {
      "get": {
        "operationId": "getSteward",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Steward"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get a steward",
        "tags": [
          "Data-1: Stewards"
        ]
      }
    },
    "/stewards/{stewardKey}/models": {
      "get": {
        "description": "Use wildcard \u0027*\u0027 for the steward key to return all models",
        "operationId": "getStewardModels",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Model"
                  }
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "Get all models from a steward",
        "tags": [
          "Data-2: Models"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}": {
      "get": {
        "operationId": "getModel",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get a model",
        "tags": [
          "Data-2: Models"
        ]
      }
    },
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionKey}": {
      "get": {
        "operationId": "getVersion",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get a version",
        "tags": [
          "Data-3: Versions"
        ]
      }
    },
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/types": {
      "get": {
        "operationId": "getAllTypes",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Type"
                  }
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get all types matching the given parameters.  Pending pagination implementation.",
        "tags": [
          "Data-6: Types"
        ]
      }
    },
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/types.cmf/{qname}": {
      "get": {
        "operationId": "getTypeCmf",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "qname",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mediaType",
            "required": false,
            "schema": {
              "type": "string",
              "default": "json",
              "enum": [
                "json",
                "xml"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get the type with the given qualified name",
        "tags": [
          "Data-6: Types"
        ]
      }
    },
...
    "/stewards/{stewardKey}/models/{modelKey}/versions/{versionNumber}/types/{qname}": {
      "get": {
        "operationId": "getType",
        "parameters": [
          {
            "in": "path",
            "name": "stewardKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "modelKey",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "versionNumber",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "qname",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Type"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get the type with the given qualified name",
        "tags": [
          "Data-6: Types"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Model": {
        "type": "object",
        "description": "A reference or message data model.",
        "properties": {
          "@id": {
            "type": "string",
            "description": "A unique identifier.  For a model, this is combines the stewardKey and modelKey fields.",
            "example": "niem/crash-driver"
          },
          "@type": {
            "type": "string",
            "description": "A kind of NIEM entity, such as a Namespace or a Property.",
            "example": "Model"
          },
          "category": {
            "type": "string",
            "description": "A kind of data model.",
            "example": "message",
            "enum": [
              "reference",
              "message",
              "other"
            ]
          },
          "description": {
            "type": "string",
            "description": "A description of a model.",
            "example": "A Crash Driver IEPD used in training to demonstrate NIEM concepts."
          },
          "developer": {
            "type": "string",
            "description": "A name of an organization or person that developed a model.",
            "example": "NIEM staff"
          },
          "fullName": {
            "type": "string",
            "description": "A full name used to identify a model.",
            "example": "NIEM Training Crash Driver IEPD"
          },
          "keywords": {
            "type": "string",
            "description": "A common alias, term, or phrase that would help to facilitate search and\n discovery of this IEPD.",
            "example": "crash, accident, vehicle, injury, charge, driver"
          },
          "localIdentifier": {
            "type": "string",
            "description": "An identifier, unique within its immediate scope.  For a model, this is the same as the modelKey field (unique within its steward).",
            "example": "crash-driver"
          },
          "modelKey": {
            "type": "string",
            "description": "A human-readable and url-friendly model identifier generated from the shortName\n field, unique within the set of models owned by its steward.",
            "example": "crash-driver"
          },
          "objective": {
            "type": "string",
            "description": "An intended use for a model, such as for actual implementation or for\n testing tool features or IEPD / message specification features.",
            "example": "example",
            "enum": [
              "implementation",
              "example",
              "test",
              "other"
            ]
          },
          "purpose": {
            "type": "string",
            "description": "A description of the intended usage and reason for which an IEPD exists.",
            "example": "This IEPD was developed to demonstrate NIEM concepts such as associations, roles, augmentations, metadata, adapters, and external standards."
          },
          "repo": {
            "type": "string",
            "description": "A code repository which hosts full IEPDs or message specifications representing the model.",
            "example": "https://github.com/niem/niem-training"
          },
          "route": {
            "type": "string",
            "description": "An endpoint to get information about a model.",
            "example": "https://tools.niem.gov/api/v2/stewards/niem/models/crash-driver"
          },
          "shortName": {
            "type": "string",
            "description": "A short name or acronym used to identify a model.",
            "example": "Crash Driver"
          },
          "steward": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "A steward responsible for the management of the model."
          },
          "subjects": {
            "type": "string",
            "description": "A description of the environment or NIEM Domain in which this IEPD is\n applicable or used.",
            "example": "Justice, Transportation, Highway Safety"
          },
          "title": {
            "type": "string",
            "description": "A steward short name and model short name.",
            "example": "NIEM Crash Driver"
          },
          "website": {
            "type": "string",
            "description": "A website address by which a person or organization may be contacted.",
            "example": "https://niem.github.io/training/"
          }
        },
        "required": [
          "shortName"
        ]
      },
...
      "Steward": {
        "type": "object",
        "description": "A program, group, or other entity responsible for managing NIEM content.",
        "properties": {
          "@id": {
            "type": "string",
            "description": "A unique identifier.  For a steward, this is the the same as the stewardKey field.",
            "example": "niem"
          },
          "@type": {
            "type": "string",
            "description": "A kind of NIEM entity, such as a Namespace or a Property.",
            "example": "Steward"
          },
          "address": {
            "type": "string",
            "description": "A complete address.",
            "example": "Hampton Roads, Virginia"
          },
          "category": {
            "type": "string",
            "description": "A kind of steward.",
            "enum": [
              "Federal",
              "State",
              "Local",
              "Tribal",
              "Territorial",
              "International",
              "Industry",
              "Nonprofit",
              "SDO",
              "Educational",
              "Person",
              "Other"
            ]
          },
          "contactName": {
            "type": "string",
            "description": "A name of the person to be contacted.",
            "example": "George P. Burdell"
          },
          "country": {
            "type": "string",
            "description": "A country, territory, dependency, or other such geopolitical subdivision of a\n location.",
            "enum": [
              "USA",
              "ABW",
              "AFG",
              "AGO",
              "AIA",
              "ALB",
              "AND",
              "ARE",
              "ARG",
              "ARM",
              "ASM",
              "ATA",
              "ATF",
              "ATG",
              "AUS",
              "AUT",
              "AX1",
              "AX2",
              "AX3",
              "AZE",
              "BDI",
              "BEL",
              "BEN",
              "BES",
              "BFA",
              "BGD",
              "BGR",
              "BHR",
              "BHS",
              "BIH",
              "BLM",
              "BLR",
              "BLZ",
              "BMU",
              "BOL",
              "BRA",
              "BRB",
              "BRN",
              "BTN",
              "BVT",
              "BWA",
              "CAF",
              "CAN",
              "CCK",
              "CHE",
              "CHL",
              "CHN",
              "CIV",
              "CMR",
              "COD",
              "COG",
              "COK",
              "COL",
              "COM",
              "CPT",
              "CPV",
              "CRI",
              "CUB",
              "CUW",
              "CXR",
              "CYM",
              "CYP",
              "CZE",
              "DEU",
              "DGA",
              "DJI",
              "DMA",
              "DNK",
              "DOM",
              "DZA",
              "ECU",
              "EGY",
              "ERI",
              "ESH",
              "ESP",
              "EST",
              "ETH",
              "FIN",
              "FJI",
              "FLK",
              "FRA",
              "FRO",
              "FSM",
              "GAB",
              "GBR",
              "GEO",
              "GGY",
              "GHA",
              "GIB",
              "GIN",
              "GLP",
              "GMB",
              "GNB",
              "GNQ",
              "GRC",
              "GRD",
              "GRL",
              "GTM",
              "GUF",
              "GUM",
              "GUY",
              "HKG",
              "HMD",
              "HND",
              "HRV",
              "HTI",
              "HUN",
              "IDN",
              "IMN",
              "IND",
              "IOT",
              "IRL",
              "IRN",
              "IRQ",
              "ISL",
              "ISR",
              "ITA",
              "JAM",
              "JEY",
              "JOR",
              "JPN",
              "KAZ",
              "KEN",
              "KGZ",
              "KHM",
              "KIR",
              "KNA",
              "KOR",
              "KWT",
              "LAO",
              "LBN",
              "LBR",
              "LBY",
              "LCA",
              "LIE",
              "LKA",
              "LSO",
              "LTU",
              "LUX",
              "LVA",
              "MAC",
              "MAF",
              "MAR",
              "MCO",
              "MDA",
              "MDG",
              "MDV",
              "MEX",
              "MHL",
              "MKD",
              "MLI",
              "MLT",
              "MMR",
              "MNE",
              "MNG",
              "MNP",
              "MOZ",
              "MRT",
              "MSR",
              "MTQ",
              "MUS",
              "MWI",
              "MYS",
              "MYT",
              "NAM",
              "NCL",
              "NER",
              "NFK",
              "NGA",
              "NIC",
              "NIU",
              "NLD",
              "NOR",
              "NPL",
              "NRU",
              "NZL",
              "OMN",
              "PAK",
              "PAN",
              "PCN",
              "PER",
              "PHL",
              "PLW",
              "PNG",
              "POL",
              "PRI",
              "PRK",
              "PRT",
              "PRY",
              "PYF",
              "QAT",
              "REU",
              "ROU",
              "RUS",
              "RWA",
              "SAU",
              "SDN",
              "SEN",
              "SGP",
              "SGS",
              "SHN",
              "SLB",
              "SLE",
              "SLV",
              "SMR",
              "SOM",
              "SPM",
              "SRB",
              "SSD",
              "STP",
              "SUR",
              "SVK",
              "SVN",
              "SWE",
              "SWZ",
              "SXM",
              "SYC",
              "SYR",
              "TCA",
              "TCD",
              "TGO",
              "THA",
              "TJK",
              "TKL",
              "TKM",
              "TLS",
              "TON",
              "TTO",
              "TUN",
              "TUR",
              "TUV",
              "TWN",
              "TZA",
              "UGA",
              "UKR",
              "URY",
              "UZB",
              "VAT",
              "VCT",
              "VEN",
              "VGB",
              "VIR",
              "VNM",
              "VUT",
              "WLF",
              "WSM",
              "XAC",
              "XAZ",
              "XBI",
              "XBK",
              "XCR",
              "XCS",
              "XCY",
              "XEU",
              "XGL",
              "XGZ",
              "XHO",
              "XJA",
              "XJM",
              "XJN",
              "XJV",
              "XKM",
              "XKN",
              "XKR",
              "XKS",
              "XMW",
              "XNV",
              "XPL",
              "XPR",
              "XQZ",
              "XSP",
              "XSV",
              "XTR",
              "XWB",
              "XWK",
              "XXD",
              "YEM",
              "ZAF",
              "ZMB",
              "ZWE"
            ]
          },
          "description": {
            "type": "string",
            "description": "A description of an organization.",
            "example": "A community-driven, standards-based approach to defining information exchange packages for multiple business domains."
          },
          "email": {
            "type": "string",
            "description": "An electronic mailing address by which a person or organization may be\n contacted.",
            "example": "information@niem.gov"
          },
          "fullName": {
            "type": "string",
            "description": "A name of an organization.",
            "example": "NIEM Management Office"
          },
          "localIdentifier": {
            "type": "string",
            "description": "An identifier, unique within its immediate scope.  For a steward, this is the same as the stewardKey field.",
            "example": "niem"
          },
          "phone": {
            "type": "string",
            "description": "A complete telephone number.",
            "example": "555-867-5309"
          },
          "route": {
            "type": "string",
            "description": "An endpoint to get information about a steward.",
            "example": "https://tools.niem.gov/api/v2/stewards/niem"
          },
          "shortName": {
            "type": "string",
            "description": "A short name or acronym used to identify a steward.  This could be the name of an\n organization or unit, a program name, or other kind of authoritative source.",
            "example": "NIEM"
          },
          "stewardKey": {
            "type": "string",
            "description": "A human-readable and url-friendly steward identifier generated from the shortName field.",
            "example": "niem"
          },
          "subunit": {
            "type": "string",
            "description": "A name of a subdivision of an organization.",
            "example": "Data and Standards Division"
          },
          "title": {
            "type": "string",
            "description": "A description used to identify an entity.",
            "example": "NIEM"
          },
          "unit": {
            "type": "string",
            "description": "A name of a high-level division of an organization.",
            "example": "Joint Staff J6"
          },
          "website": {
            "type": "string",
            "description": "A website address by which a person or organization may be contacted.",
            "example": "https://niem.gov"
          }
        },
        "required": [
          "shortName"
        ]
      },
...
    }
  }
}